+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Use `failure` instead of `error-chain`.
+- Minimum Rust version is 1.18.
+- New error messages.
+- `TokenType` is properly public now.
+
+### Removed
+- `ChainedError`
 
 ## [0.2.0] - 2018-03-11
 ### Added